Fix Jacoco coverage repots in Sonar (#136)
Now that inject and plexus are consolidated, they should not step on each other toe, and write to top level target directory the report.

This does NOT fix Sonar, that for me, unknown reason, does not pick up generated Jacoco reports. Still, the reports on local builds are there and are usable.